feat(default-baseline): promote parquet runner to canonical
Promote the parquet sibling (scripts/run_default_baseline_parquet.py)
to the canonical scripts/run_default_baseline.py, replacing the legacy
sqlite runner. The workspace-default emitter
(workspace.yaml.runtime.default_emitter = parquet) is now what captures
the reference state every new study falls back to — no more two-runner
forking decision at the workspace level.

Upstream blockers from the 2026-05-26 docstring are now resolved:
  * pbg-emitters PR #82 fixed the multi-dim ndarray polars "dtype('O')"
    wall — real cell composites now write parquet cleanly.
  * vivarium-dashboard PR #136 surfaces ParquetEmitter runs in
    /api/simulations so both shapes co-exist on the Simulations DB tab.

What changed:
  * scripts/run_default_baseline.py — replaced (the legacy sqlite content
    is deleted; the parquet content from the sibling now sits under this
    canonical name).
  * scripts/run_default_baseline_parquet.py — removed (consolidated).
  * Docstring updated: removed stale "real-composite status" + "dashboard
    interop note" warnings, added promotion note.

Smoke-tested: 1-step (60 sim-sec) run wrote 5.9MB of parquet under
.pbg/default-baseline/parquet/<experiment_id>/history/...; structure
matches the per-study hive shape the dashboard already discovers.
