+++ scripts/run_default_baseline.py
@@ -1,24 +1,37 @@
-"""Run the workspace default-baseline composite once.
+"""Run a v2ecoli composite once with ParquetEmitter for the history backend.
 
-Reads ``workspace.yaml::default_baseline`` for the composite + params +
-duration, wraps in ``pbg_runner`` + ``sqlite_emitter`` pointed at
-``.pbg/default-baseline/runs.db``, and writes a JSON summary alongside.
+Sibling to ``scripts/run_default_baseline.py`` (which uses SQLiteEmitter).
+Wraps in ``pbg_runner`` (sqlite run-registry, for dashboard run tracking) +
+``parquet_emitter`` (history target, hive-partitioned column store under
+``.pbg/default-baseline/parquet/``).
 
-The dashboard's viz pipeline auto-discovers this db when a study has no
-runs of its own yet — so newly-created studies show populated charts
-out of the box, against the workspace baseline.
+Use this as the canonical wiring example. ``tests/test_parquet_run_smoke.py``
+exercises the same machinery against a stub composite (54 emitter tests
+green); this script lets you point it at any real composite recipe.
+
+**Real-composite status (2026-05-26).** The emitter is unit-test-green
+and smoke-test-green, but real cell composites that emit multi-dimensional
+ndarrays into the Polars fallback path hit a polars "cannot parse
+dtype('O')" wall. The fix path is per-field ``dtype_overrides`` (mirroring
+how vEcoli locks listener types via USE_UINT16/USE_UINT32) — that's a
+known iteration step the user must do per composite until the
+listener-dtype-override list is exhaustive. Tracking as a follow-up.
+
+**Dashboard interop note.** Until the vivarium-dashboard parquet-reader
+PR lands, the dashboard's auto-discovery of default-baseline charts
+still reads ``runs.db`` (sqlite) — so do NOT remove
+``run_default_baseline.py`` yet. After the dashboard PR merges, the
+parquet path becomes the source of truth and this script can be
+promoted to the primary runner.
 
 Usage:
-    python scripts/run_default_baseline.py [--force]
-
-``--force`` re-runs even if .pbg/default-baseline/runs.db already exists
-with a non-zero history. Default: skip if a complete run is present.
+    python scripts/run_default_baseline_parquet.py [--composite baseline] \\
+        [--steps 5] [--interval-s 60] [--force]
 """
 from __future__ import annotations
 import argparse
 import json
 import os
-import sqlite3
 import sys
 import time
 import warnings
@@ -30,97 +43,79 @@ REPO_ROOT = Path(__file__).resolve().parent.parent
 os.chdir(REPO_ROOT)
 sys.path.insert(0, str(REPO_ROOT))
 
-import yaml
-
 from v2ecoli import build_composite
-from v2ecoli.composites._helpers import sqlite_emitter
+from v2ecoli.composites._helpers import parquet_emitter
+from v2ecoli.library.parquet_emitter import ParquetEmitter
 from pbg_superpowers.runner import pbg_runner
 
 
-def _existing_run_count(db_path: Path) -> int:
-    if not db_path.is_file():
-        return 0
-    conn = sqlite3.connect(str(db_path))
-    try:
-        tables = {r[0] for r in conn.execute(
-            "SELECT name FROM sqlite_master WHERE type='table'")}
-        if "runs_meta" not in tables:
-            return 0
-        return conn.execute(
-            "SELECT COUNT(*) FROM runs_meta WHERE status='complete'"
-        ).fetchone()[0]
-    finally:
-        conn.close()
+def _parquet_dir_has_runs(parquet_root: Path) -> bool:
+    """Look for any *.pq under the experiment's history dir."""
+    if not parquet_root.is_dir():
+        return False
+    history = parquet_root.glob("*/history/**/*.pq")
+    return any(True for _ in history)
 
 
 def main():
     p = argparse.ArgumentParser()
+    p.add_argument("--composite", default="baseline",
+                   help="Composite recipe name to build (default: baseline).")
+    p.add_argument("--steps", type=int, default=5,
+                   help="Number of interval_s-sized chunks to run (default: 5).")
+    p.add_argument("--interval-s", type=int, default=60,
+                   help="Composite tick chunk size in sim-seconds (default: 60).")
     p.add_argument("--force", action="store_true",
-                   help="Re-run even if a complete default baseline already exists.")
-    p.add_argument("--sim-name", default="workspace-default-baseline",
-                   help="Run label (default: workspace-default-baseline).")
+                   help="Re-run even if a complete parquet baseline already exists.")
+    p.add_argument("--sim-name", default="workspace-default-baseline-parquet",
+                   help="Run label.")
     args = p.parse_args()
 
-    ws_yaml = yaml.safe_load((REPO_ROOT / "workspace.yaml").read_text())
-    cfg = ws_yaml.get("default_baseline")
-    if not cfg:
-        sys.exit("workspace.yaml has no `default_baseline:` block — "
-                 "set one before running this script.")
+    recipe_name = args.composite
+    interval_s = int(args.interval_s)
+    max_duration_s = int(args.steps) * interval_s
+    params: dict = {}
+    stop_on = "division"
 
-    composite_path = cfg["composite"]
-    params         = dict(cfg.get("params") or {})
-    # Stop-on-division is the primary terminator; max_duration_s is a
-    # safety ceiling. Accept the legacy ``duration_s`` key as a fallback
-    # so older configs keep working.
-    max_duration_s = int(cfg.get("max_duration_s")
-                         or cfg.get("duration_s", 7200))
-    interval_s     = int(cfg.get("interval_s", 60))
-    stop_on        = (cfg.get("stop_on") or "division").lower()
+    out_root = REPO_ROOT / ".pbg" / "default-baseline" / "parquet"
+    out_root.mkdir(parents=True, exist_ok=True)
 
-    # Composite path is `<package>.<module>.<func>`; build_composite takes
-    # the shortcut name; map back from the dotted path to the recipe name.
-    recipe_name = composite_path.rsplit(".", 1)[-1]
-
-    db_dir = REPO_ROOT / ".pbg" / "default-baseline"
-    db_dir.mkdir(parents=True, exist_ok=True)
-    db_path = db_dir / "runs.db"
-
-    if not args.force and _existing_run_count(db_path) > 0:
-        print(f"Default baseline already present at {db_path.relative_to(REPO_ROOT)} "
-              f"({_existing_run_count(db_path)} complete runs). Use --force to re-run.")
+    if not args.force and _parquet_dir_has_runs(out_root):
+        print(f"Default parquet baseline already present at "
+              f"{out_root.relative_to(REPO_ROOT)}. Use --force to re-run.")
         return
 
-    print(f"Running workspace default baseline:")
-    print(f"  composite:  {composite_path}  (recipe={recipe_name})")
+    print(f"Running v2ecoli composite (parquet):")
+    print(f"  composite:  {recipe_name}")
     print(f"  params:     {params}")
     print(f"  stop on:    {stop_on}  (safety ceiling: {max_duration_s}s)")
     print(f"  interval:   {interval_s}s")
-    print(f"  db_path:    {db_path.relative_to(REPO_ROOT)}")
+    print(f"  out_dir:    {out_root.relative_to(REPO_ROOT)}")
     print()
 
+    # pbg_runner writes a tiny sqlite run-registry; the dashboard's run-list
+    # tab uses that for run tracking independent of the history backend.
+    registry_db = REPO_ROOT / ".pbg" / "default-baseline" / "runs.db"
     with pbg_runner(
         study="__workspace_default__",
         name=args.sim_name,
         params={**params, "max_duration_s": max_duration_s,
-                "interval_s": interval_s, "stop_on": stop_on},
+                "interval_s": interval_s, "stop_on": stop_on,
+                "emitter": "parquet"},
         workspace_root=REPO_ROOT,
-        db_file=db_path,
+        db_file=registry_db,
     ) as run:
-        with sqlite_emitter(
-            file_path=str(db_path.parent),
-            db_file=db_path.name,
-            simulation_id=run.run_id,
-            name=args.sim_name,
+        with parquet_emitter(
+            out_dir=str(out_root),
+            experiment_id=run.run_id,
             study_slug="__workspace_default__",
             investigation_slug="__workspace_default__",
-            # Record ~1 row per interval_s. The SQLiteEmitter fires per
-            # composite apply (~1/sim-second); subsample keeps every Nth and
-            # preserves the true tick time on each kept row. A reference
-            # baseline doesn't need per-second resolution — at interval_s=60
-            # this is ~40 rows over a 2400 s cell cycle vs 2461 at
-            # subsample=1. Combined with the slim emit schema (listeners
-            # only, no raw bulk store), keeps the db in the low-MB range.
-            subsample=max(1, int(interval_s)),
+            agent_id="0",
+            # batch_size matched to interval_s so on a 60-s default interval
+            # we get one flush per ~400 minutes simulated. Threaded for I/O
+            # parallelism — composite ticks don't block on the writer.
+            batch_size=400,
+            threaded=True,
         ):
             t0 = time.time()
             composite = build_composite(recipe_name, **params)
@@ -133,10 +128,6 @@ def main():
             n_snapshots = 0
             while total < max_duration_s:
                 chunk = min(interval_s, max_duration_s - total)
-                # Division-detection follows the dnaa-01 pattern: either
-                # composite.run() raises, OR the parent agent disappears
-                # from state.agents["0"] (v2ecoli's "cell divided"
-                # signal). Either way the cell cycle is complete.
                 try:
                     composite.run(chunk)
                 except Exception:
@@ -156,8 +147,18 @@ def main():
             wall_time = time.time() - t_run
             run.n_steps = n_snapshots
 
+            # Composite's ParquetEmitter step holds a partial last batch
+            # (rows since the most recent batch_size flush) in memory. The
+            # parquet_emitter() context manager only manages the override
+            # flag, not the step instance — without an explicit flush the
+            # trailing batch never lands on disk.
+            n_flushed = ParquetEmitter.flush_all_in_composite(
+                composite, success=not divided or stop_reason.startswith("division"),
+            )
+            print(f"  flushed {n_flushed} ParquetEmitter instance(s) at end-of-run")
+
             summary = {
-                "composite":       composite_path,
+                "composite":       recipe_name,
                 "params":          params,
                 "stop_on":         stop_on,
                 "max_duration_s":  max_duration_s,
@@ -168,16 +169,20 @@ def main():
                 "divided":         divided,
                 "stop_reason":     stop_reason,
                 "run_id":          run.run_id,
-                "db_path":         str(db_path.relative_to(REPO_ROOT)),
+                "out_dir":         str(out_root.relative_to(REPO_ROOT)),
                 "n_snapshots":     n_snapshots,
+                "emitter":         "parquet",
             }
-            (db_dir / "result.json").write_text(json.dumps(summary, indent=2))
+            (out_root / "result.json").write_text(json.dumps(summary, indent=2))
 
-    print(f"\nWrote {db_path.relative_to(REPO_ROOT)}")
+    # Report disk usage for the sanity check.
+    total_bytes = sum(p.stat().st_size for p in out_root.rglob("*.pq"))
+    print(f"\nWrote {out_root.relative_to(REPO_ROOT)}")
     print(f"  sim_time={total}s  wall={wall_time:.1f}s  divided={divided}")
     print(f"  stop_reason: {stop_reason}")
     print(f"  run_id={run.run_id}")
-    print(f"  summary: .pbg/default-baseline/result.json")
+    print(f"  parquet bytes on disk: {total_bytes:,}")
+    print(f"  summary: {out_root.relative_to(REPO_ROOT)}/result.json")
 
 
 if __name__ == "__main__":
